storjscandb: create database (#24)
* storjscandb: create database

Create storjscan database and db testing pkg. Database is created
using dbx and contains block header model which is used by block
header cache that is introduced in the following changes and 1 migration
which creates table for block headers. The purpose of this change is to
unblock other changes which requires db interaction and to prevent
reimplementing common db initialisation code in that changes.

* disable cockroach

* remove references to cockroach and add simple testdb run test

wip

wip

wip