tokenprice: replace transaction "upsert" with insert on conflict update
The transaction in the update method of the priceQuoteDB is not needed as it does not provide the locking required to handle multiple updates at once. Though it is unlikely to have such multiple updates, to avoid a race condition, the best approach is to use an "upsert" which will attempt an insert, and update instead on conflict.

Change-Id: I43a828c9afff17f780045af4e78cc36a3a7c54d5