dockerfile: add integration test
Adds a docker file that can be used for local testing. Also adds a script that can be run via mage to test basic endpoint functionality by spinning up the services in docker, and curling the standard endpoints. No verification of the endpoints is currently done as a part of the test. The magefile also provides helper methods for building and pushing the dockerfile, as well as code coverage, linter and a few other helpful commands.

Change-Id: I73436496d173cfe9e87b12129afa9b7f436eb9b6