{blockchain/events,tokens}: replace events cache and chore with service
Remove the use of the events cache, and replace it with an events
service. Instead of storing the events in the database as the cache did,
the service will keep a local map with the last scanned block number.
Using this map, allows for the service to only scan blocks that is has
not seen and confirmed. If the service is restarted, it will fall back
and scan blocks based on the last seen transactions within the satellite.

Additionally, the service will be triggered to perform a new scan for
events when requested by the client(satellite), rather than using an
internal chore. Any found events are returned to the client(satellite)
directly.

Change-Id: I65bfe9b71cc145ca310f03200a5080cd78ee9dfa